requirements.txt
streamlit==1.39.0
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
huggingface_hub==0.26.0
sentence-transformers==3.2.0
faiss-cpu==1.9.0
python-dotenv==1.0.1